+++ packages/core/src/render3/instructions/animation.ts
@@ -6,143 +6,40 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {stringify} from '../../util/stringify'; // Adjust imports as per actual location
 import {
   AnimationCallbackEvent,
   AnimationFunction,
-  ANIMATIONS_DISABLED,
   MAX_ANIMATION_TIMEOUT,
-  LongestAnimation,
 } from '../../animation/interfaces';
 import {getLView, getCurrentTNode} from '../state';
-import {
-  RENDERER,
-  INJECTOR,
-  CONTEXT,
-  LView,
-  DECLARATION_LCONTAINER,
-  ANIMATIONS,
-} from '../interfaces/view';
-import {RuntimeError, RuntimeErrorCode} from '../../errors';
+import {RENDERER, INJECTOR, CONTEXT, LView} from '../interfaces/view';
 import {getNativeByTNode} from '../util/view_utils';
 import {performanceMarkFeature} from '../../util/performance';
 import {Renderer} from '../interfaces/renderer';
-import {RElement} from '../interfaces/renderer_dom';
 import {NgZone} from '../../zone';
 import {determineLongestAnimation, allLeavingAnimations} from '../../animation/longest_animation';
 import {TNode} from '../interfaces/node';
-import {getBeforeNodeForView} from '../node_manipulation';
 import type {PromiseConstructor} from '../../util/promise_with_resolvers';
 
-const DEFAULT_ANIMATIONS_DISABLED = false;
-const areAnimationSupported =
-  (typeof ngServerMode === 'undefined' || !ngServerMode) &&
-  typeof document !== 'undefined' &&
-  // tslint:disable-next-line:no-toplevel-property-access
-  typeof document?.documentElement?.getAnimations === 'function';
-
-/**
- * Helper function to check if animations are disabled via injection token
- */
-function areAnimationsDisabled(lView: LView): boolean {
-  const injector = lView[INJECTOR]!;
-  return injector.get(ANIMATIONS_DISABLED, DEFAULT_ANIMATIONS_DISABLED);
-}
-
-/**
- * Helper function to cleanup enterClassMap data safely
- */
-function cleanupEnterClassData(element: HTMLElement): void {
-  const elementData = enterClassMap.get(element);
-  if (elementData) {
-    for (const fn of elementData.cleanupFns) {
-      fn();
-    }
-    enterClassMap.delete(element);
-  }
-  longestAnimations.delete(element);
-}
-
-const noOpAnimationComplete = () => {};
-
-// Tracks the list of classes added to a DOM node from `animate.enter` calls to ensure
-// we remove all of the classes in the case of animation composition via host bindings.
-const enterClassMap = new WeakMap<HTMLElement, {classList: string[]; cleanupFns: Function[]}>();
-const longestAnimations = new WeakMap<HTMLElement, LongestAnimation>();
-
-// Tracks nodes that are animating away for the duration of the animation. This is
-// used to prevent duplicate nodes from showing up when nodes have been toggled quickly
-// from an `@if` or `@for`.
-const leavingNodes = new WeakMap<TNode, HTMLElement[]>();
-
-function clearLeavingNodes(tNode: TNode, el: HTMLElement): void {
-  const nodes = leavingNodes.get(tNode);
-  if (nodes && nodes.length > 0) {
-    const ix = nodes.findIndex((node) => node === el);
-    if (ix > -1) nodes.splice(ix, 1);
-  }
-  if (nodes?.length === 0) {
-    leavingNodes.delete(tNode);
-  }
-}
-
-/**
- * In the case that we have an existing node that's animating away, like when
- * an `@if` toggles quickly, we need to end the animation for the former node
- * and remove it right away to prevent duplicate nodes showing up.
- */
-function cancelLeavingNodes(tNode: TNode, lView: LView): void {
-  const leavingEl = leavingNodes.get(tNode)?.shift();
-  const lContainer = lView[DECLARATION_LCONTAINER];
-  if (lContainer) {
-    // this is the insertion point for the new TNode element.
-    // it will be inserted before the declaring containers anchor.
-    const beforeNode = getBeforeNodeForView(tNode.index, lContainer);
-    // here we need to check the previous sibling of that anchor. The first
-    // previousSibling node will be the new element added. The second
-    // previousSibling will be the one that's being removed.
-    const previousNode = beforeNode?.previousSibling;
-    // We really only want to cancel animations if the leaving node is the
-    // same as the node before where the new node will be inserted. This is
-    // the control flow scenario where an if was toggled.
-    if (leavingEl && previousNode && leavingEl === previousNode) {
-      leavingEl.dispatchEvent(new CustomEvent('animationend', {detail: {cancel: true}}));
-    }
-  }
-}
-
-function trackLeavingNodes(tNode: TNode, el: HTMLElement): void {
-  // We need to track this tNode's element just to be sure we don't add
-  // a new RNode for this TNode while this one is still animating away.
-  // once the animation is complete, we remove this reference.
-  if (leavingNodes.has(tNode)) {
-    leavingNodes.get(tNode)?.push(el);
-  } else {
-    leavingNodes.set(tNode, [el]);
-  }
-}
-
-function getLViewEnterAnimations(lView: LView): Function[] {
-  const animationData = (lView[ANIMATIONS] ??= {});
-  return (animationData.enter ??= []);
-}
-
-function getLViewLeaveAnimations(lView: LView): Function[] {
-  const animationData = (lView[ANIMATIONS] ??= {});
-  return (animationData.leave ??= []);
-}
-
-function getClassListFromValue(value: string | Function | string[]): string[] | null {
-  const classes = typeof value === 'function' ? value() : value;
-  let classList: string[] | null = Array.isArray(classes) ? classes : null;
-  if (typeof classes === 'string') {
-    classList = classes
-      .trim()
-      .split(/\s+/)
-      .filter((k) => k);
-  }
-  return classList;
-}
+import {
+  areAnimationsDisabled,
+  areAnimationSupported,
+  assertAnimationTypes,
+  assertElementNodes,
+  cancelAnimationsIfRunning,
+  cancelLeavingNodes,
+  cleanupEnterClassData,
+  clearLeavingNodes,
+  enterClassMap,
+  getClassListFromValue,
+  getLViewEnterAnimations,
+  getLViewLeaveAnimations,
+  isLongestAnimation,
+  longestAnimations,
+  noOpAnimationComplete,
+  trackEnterClasses,
+  trackLeavingNodes,
+} from '../../animation/utils';
 
 /**
  * Instruction to handle the `animate.enter` behavior for class bindings.
@@ -191,23 +88,23 @@ export function runEnterAnimation(lView: LView, tNode: TNode, value: string | Fu
   // to get the longest animation to ensure we don't complete animations early.
   // This also allows us to setup cancellation of animations in progress if the
   // gets removed early.
-  const handleAnimationStart = (event: AnimationEvent | TransitionEvent) => {
+  const handleEnterAnimationStart = (event: AnimationEvent | TransitionEvent) => {
     const eventName = event instanceof AnimationEvent ? 'animationend' : 'transitionend';
     ngZone.runOutsideAngular(() => {
-      cleanupFns.push(renderer.listen(nativeElement, eventName, handleInAnimationEnd));
+      cleanupFns.push(renderer.listen(nativeElement, eventName, handleEnterAnimationEnd));
     });
   };
 
   // When the longest animation ends, we can remove all the classes
-  const handleInAnimationEnd = (event: AnimationEvent | TransitionEvent) => {
-    animationEnd(event, nativeElement, renderer);
+  const handleEnterAnimationEnd = (event: AnimationEvent | TransitionEvent) => {
+    enterAnimationEnd(event, nativeElement, renderer);
   };
 
   // We only need to add these event listeners if there are actual classes to apply
   if (activeClasses && activeClasses.length > 0) {
     ngZone.runOutsideAngular(() => {
-      cleanupFns.push(renderer.listen(nativeElement, 'animationstart', handleAnimationStart));
-      cleanupFns.push(renderer.listen(nativeElement, 'transitionstart', handleAnimationStart));
+      cleanupFns.push(renderer.listen(nativeElement, 'animationstart', handleEnterAnimationStart));
+      cleanupFns.push(renderer.listen(nativeElement, 'transitionstart', handleEnterAnimationStart));
     });
 
     trackEnterClasses(nativeElement, activeClasses, cleanupFns);
@@ -232,23 +129,23 @@ export function runEnterAnimation(lView: LView, tNode: TNode, value: string | Fu
   }
 }
 
-/**
- * trackEnterClasses is necessary in the case of composition where animate.enter
- * is used on the same element in multiple places, like on the element and in a
- * host binding. When removing classes, we need the entire list of animation classes
- * added to properly remove them when the longest animation fires.
- */
-function trackEnterClasses(el: HTMLElement, classList: string[], cleanupFns: Function[]) {
-  const elementData = enterClassMap.get(el);
-  if (elementData) {
-    for (const klass of classList) {
-      elementData.classList.push(klass);
+function enterAnimationEnd(
+  event: AnimationEvent | TransitionEvent,
+  nativeElement: HTMLElement,
+  renderer: Renderer,
+) {
+  const elementData = enterClassMap.get(nativeElement);
+  if (!elementData) return;
+  if (isLongestAnimation(event, nativeElement)) {
+    // Now that we've found the longest animation, there's no need
+    // to keep bubbling up this event as it's not going to apply to
+    // other elements further up. We don't want it to inadvertently
+    // affect any other animations on the page.
+    event.stopImmediatePropagation();
+    for (const klass of elementData.classList) {
+      renderer.removeClass(nativeElement, klass);
     }
-    for (const fn of cleanupFns) {
-      elementData.cleanupFns.push(fn);
-    }
-  } else {
-    enterClassMap.set(el, {classList, cleanupFns});
+    cleanupEnterClassData(nativeElement);
   }
 }
 
@@ -361,6 +258,71 @@ function runLeaveAnimations(
 }
 
 /**
+ * This function actually adds the classes that animate element that's leaving the DOM.
+ * Once it finishes, it calls the remove function that was provided by the DOM renderer.
+ */
+function animateLeaveClassRunner(
+  el: HTMLElement,
+  tNode: TNode,
+  classList: string[],
+  renderer: Renderer,
+  animationsDisabled: boolean,
+  ngZone: NgZone,
+  resolver: VoidFunction,
+) {
+  if (animationsDisabled) {
+    longestAnimations.delete(el);
+    resolver();
+    return;
+  }
+
+  cancelAnimationsIfRunning(el, renderer);
+
+  const handleOutAnimationEnd = (event: AnimationEvent | TransitionEvent | CustomEvent) => {
+    if (event instanceof CustomEvent || isLongestAnimation(event, el)) {
+      // Now that we've found the longest animation, there's no need
+      // to keep bubbling up this event as it's not going to apply to
+      // other elements further up. We don't want it to inadvertently
+      // affect any other animations on the page.
+      event.stopImmediatePropagation();
+      longestAnimations.delete(el);
+      clearLeavingNodes(tNode, el);
+
+      if (Array.isArray(tNode.projection)) {
+        // in the content projection case, the element is not destroyed.
+        // So we need to remove the class at the end so that it isn't left
+        // behind for whenever the item shows up again.
+        for (const item of classList) {
+          renderer.removeClass(el, item);
+        }
+      }
+    }
+    resolver();
+  };
+
+  ngZone.runOutsideAngular(() => {
+    renderer.listen(el, 'animationend', handleOutAnimationEnd);
+    renderer.listen(el, 'transitionend', handleOutAnimationEnd);
+  });
+  trackLeavingNodes(tNode, el);
+  for (const item of classList) {
+    renderer.addClass(el, item);
+  }
+  // In the case that the classes added have no animations, we need to remove
+  // the element right away. This could happen because someone is intentionally
+  // preventing an animation via selector specificity.
+  ngZone.runOutsideAngular(() => {
+    requestAnimationFrame(() => {
+      determineLongestAnimation(el, longestAnimations, areAnimationSupported);
+      if (!longestAnimations.has(el)) {
+        clearLeavingNodes(tNode, el);
+        resolver();
+      }
+    });
+  });
+}
+
+/**
  * Instruction to handle the `(animate.leave)` behavior for event bindings, aka when
  * a user wants to use a custom animation function rather than a class. It registers
  * a leave animation function in the LView to be run at right before removal from the
@@ -444,145 +406,4 @@ function runLeaveAnimationFunction(
 
   // Ensure cleanup if the LView is destroyed before the animation runs.
   return promise;
-}
-
-function cancelAnimationsIfRunning(element: HTMLElement, renderer: Renderer): void {
-  if (!areAnimationSupported) return;
-  const elementData = enterClassMap.get(element);
-  if (
-    elementData &&
-    elementData.classList.length > 0 &&
-    elementHasClassList(element, elementData.classList)
-  ) {
-    for (const klass of elementData.classList) {
-      renderer.removeClass(element as unknown as RElement, klass);
-    }
-  }
-  // We need to prevent any enter animation listeners from firing if they exist.
-  cleanupEnterClassData(element);
-}
-
-function elementHasClassList(element: HTMLElement, classList: string[]): boolean {
-  for (const className of classList) {
-    if (element.classList.contains(className)) return true;
-  }
-  return false;
-}
-
-function isLongestAnimation(
-  event: AnimationEvent | TransitionEvent,
-  nativeElement: HTMLElement,
-): boolean {
-  const longestAnimation = longestAnimations.get(nativeElement);
-  return (
-    nativeElement === event.target &&
-    longestAnimation !== undefined &&
-    ((longestAnimation.animationName !== undefined &&
-      (event as AnimationEvent).animationName === longestAnimation.animationName) ||
-      (longestAnimation.propertyName !== undefined &&
-        (event as TransitionEvent).propertyName === longestAnimation.propertyName))
-  );
-}
-
-function animationEnd(
-  event: AnimationEvent | TransitionEvent,
-  nativeElement: HTMLElement,
-  renderer: Renderer,
-) {
-  const elementData = enterClassMap.get(nativeElement);
-  if (!elementData) return;
-  if (isLongestAnimation(event, nativeElement)) {
-    // Now that we've found the longest animation, there's no need
-    // to keep bubbling up this event as it's not going to apply to
-    // other elements further up. We don't want it to inadvertently
-    // affect any other animations on the page.
-    event.stopImmediatePropagation();
-    for (const klass of elementData.classList) {
-      renderer.removeClass(nativeElement, klass);
-    }
-    cleanupEnterClassData(nativeElement);
-  }
-}
-
-function assertAnimationTypes(value: string | Function, instruction: string) {
-  if (value == null || (typeof value !== 'string' && typeof value !== 'function')) {
-    throw new RuntimeError(
-      RuntimeErrorCode.ANIMATE_INVALID_VALUE,
-      `'${instruction}' value must be a string of CSS classes or an animation function, got ${stringify(value)}`,
-    );
-  }
-}
-
-function assertElementNodes(nativeElement: Element, instruction: string) {
-  if ((nativeElement as Node).nodeType !== Node.ELEMENT_NODE) {
-    throw new RuntimeError(
-      RuntimeErrorCode.ANIMATE_INVALID_VALUE,
-      `'${instruction}' can only be used on an element node, got ${stringify((nativeElement as Node).nodeType)}`,
-    );
-  }
-}
-
-/**
- * This function actually adds the classes that animate element that's leaving the DOM.
- * Once it finishes, it calls the remove function that was provided by the DOM renderer.
- */
-function animateLeaveClassRunner(
-  el: HTMLElement,
-  tNode: TNode,
-  classList: string[],
-  renderer: Renderer,
-  animationsDisabled: boolean,
-  ngZone: NgZone,
-  resolver: VoidFunction,
-) {
-  if (animationsDisabled) {
-    longestAnimations.delete(el);
-    resolver();
-    return;
-  }
-
-  cancelAnimationsIfRunning(el, renderer);
-
-  const handleOutAnimationEnd = (event: AnimationEvent | TransitionEvent | CustomEvent) => {
-    if (event instanceof CustomEvent || isLongestAnimation(event, el)) {
-      // Now that we've found the longest animation, there's no need
-      // to keep bubbling up this event as it's not going to apply to
-      // other elements further up. We don't want it to inadvertently
-      // affect any other animations on the page.
-      event.stopImmediatePropagation();
-      longestAnimations.delete(el);
-      clearLeavingNodes(tNode, el);
-
-      if (Array.isArray(tNode.projection)) {
-        // in the content projection case, the element is not destroyed.
-        // So we need to remove the class at the end so that it isn't left
-        // behind for whenever the item shows up again.
-        for (const item of classList) {
-          renderer.removeClass(el, item);
-        }
-      }
-    }
-    resolver();
-  };
-
-  ngZone.runOutsideAngular(() => {
-    renderer.listen(el, 'animationend', handleOutAnimationEnd);
-    renderer.listen(el, 'transitionend', handleOutAnimationEnd);
-  });
-  trackLeavingNodes(tNode, el);
-  for (const item of classList) {
-    renderer.addClass(el, item);
-  }
-  // In the case that the classes added have no animations, we need to remove
-  // the element right away. This could happen because someone is intentionally
-  // preventing an animation via selector specificity.
-  ngZone.runOutsideAngular(() => {
-    requestAnimationFrame(() => {
-      determineLongestAnimation(el, longestAnimations, areAnimationSupported);
-      if (!longestAnimations.has(el)) {
-        clearLeavingNodes(tNode, el);
-        resolver();
-      }
-    });
-  });
 }
